added hex and oct literals, fixed dec literals, added 32 bit literal requirement

+++ public/vush/vush.hpp
@@ -10,7 +10,7 @@
 namespace vush {
     struct Constant_Define {
         anton::String name;
-        i64 value;
+        i32 value;
     };
 
     struct Format_Options {
